+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-6499] Add support for legacy API connections
+
 ## [6.8.1] 2024-06-17
 
 ### Fixed